fix: 정답 제출 시 문제 아이디의 타입이 맞지 않아 정답이 제출되지 않는 현상 수정

+++ packages/backend/src/quiz/entity/category-question.entity.ts
@@ -10,7 +10,7 @@ export class CategoryQuestion {
   @Column({ type: 'bigint', nullable: false, name: 'category_id' })
   categoryId: number;
 
-  @Column({ type: 'bigint', nullable: false, name: 'question_id' })
+  @Column({ type: 'int4', nullable: false, name: 'question_id' })
   questionId: number;
 
   @ManyToOne(() => Category, (category) => category.categoryQuestions)
